Refactor date functions and improve handling of DOB file
This commit overhauls the date handling functions and streamlines functions related to the date-of-birth file. Date functions now are more efficient and use the new `getCurrentDate` function for improved accuracy. The way we handle the date-of-birth file has been revamped, including changing from text to binary storage, and optimizing file existence checks and reading the saved date.