Update executable path and ignore main output file
The path for the main executable in the CMakeLists.txt file has been corrected. `main` output file is now being ignored in the .gitignore configuration file to ensure the compiled version of the program is not committed to the repository. Moreover, time calculations for an individual's age in days, hours, minutes, and seconds have been implemented in main.c.